--- src/main/java/com/daitda/hubservice/HubServiceApplication.java
@@ -1,13 +1,0 @@
-package com.daitda.hubservice;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class HubServiceApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(HubServiceApplication.class, args);
-    }
-
-}
