--- src/main/java/com/daitda/hubservice/hub/dto/request/HubUpdateRequest.java
@@ -1,33 +1,0 @@
-package com.daitda.hubservice.hub.dto.request;
-
-
-import jakarta.validation.constraints.DecimalMax;
-import jakarta.validation.constraints.DecimalMin;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
-import lombok.Getter;
-
-import java.math.BigDecimal;
-
-@Getter
-public class HubUpdateRequest {
-
-    @NotBlank(message = "허브명은 필수입니다.")
-    private String hubName;
-
-    @NotBlank(message = "허브 주소는 필수입니다.")
-    private String hubAddress;
-
-    @NotNull(message = "위도는 필수입니다.")
-    @DecimalMin(value = "-90.0", message = "위도는 -90 이상이어야 합니다.")
-    @DecimalMax(value = "90.0", message = "위도는 90 이하여야 합니다.")
-    private BigDecimal latitude;
-
-    @NotNull(message = "경도는 필수입니다.")
-    @DecimalMin(value = "-180.0", message = "경도는 -180 이상이어야 합니다.")
-    @DecimalMax(value = "180.0", message = "경도는 180 이하여야 합니다.")
-    private BigDecimal longitude;
-
-    @NotNull(message = "중앙 허브 여부는 필수입니다.")
-    private Boolean isCentral;
-}
