--- src/main/java/com/daitda/hubservice/hubinventory/application/dto/command/UpdateHubInventoryCommand.java
@@ -1,12 +1,0 @@
-package com.daitda.hubservice.hubinventory.application.dto.command;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class UpdateHubInventoryCommand {
-
-    private Integer quantity;
-
-}
