--- src/main/java/com/daitda/hubservice/hubinventory/application/result/ListHubInventoryResult.java
@@ -1,23 +1,0 @@
-package com.daitda.hubservice.hubinventory.application.result;
-
-import com.daitda.hubservice.hubinventory.domain.entity.HubInventory;
-
-import java.util.UUID;
-
-public record ListHubInventoryResult(
-        UUID hubInventoryId,
-        UUID hubId,
-        UUID companyId,
-        UUID productId,
-        Integer quantity
-) {
-    public static ListHubInventoryResult from(HubInventory hubInventory) {
-        return new ListHubInventoryResult(
-                hubInventory.getHubInventoryId(),
-                hubInventory.getHubId(),
-                hubInventory.getCompanyId(),
-                hubInventory.getProductId(),
-                hubInventory.getQuantity()
-        );
-    }
-}
