--- src/main/java/com/daitda/hubservice/hubinventory/presentation/dto/request/UpdateHubInventoryRequest.java
@@ -1,13 +1,0 @@
-package com.daitda.hubservice.hubinventory.presentation.dto.request;
-
-import jakarta.validation.constraints.Min;
-import jakarta.validation.constraints.NotNull;
-import lombok.Getter;
-
-@Getter
-public class UpdateHubInventoryRequest {
-
-    @NotNull(message = "재고 수량은 필수입니다.")
-    @Min(value = 0, message = "재고 수량은 0 이상이어야 합니다.")
-    private Integer quantity;
-}
