+++ src/main/java/com/fhsh/daitda/hubservice/HubServiceApplication.java
@@ -1,0 +1,13 @@
+package com.fhsh.daitda.hubservice;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class HubServiceApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(HubServiceApplication.class, args);
+    }
+
+}
