+++ src/main/java/com/fhsh/daitda/hubservice/hubinventory/application/dto/command/CreateHubInventoryCommand.java
@@ -1,0 +1,17 @@
+package com.fhsh.daitda.hubservice.hubinventory.application.dto.command;
+
+import lombok.Builder;
+import lombok.Getter;
+
+import java.util.UUID;
+
+@Getter
+@Builder
+public class CreateHubInventoryCommand {
+
+    private UUID hubId;
+    private UUID companyId;
+    private UUID productId;
+    private Integer quantity;
+
+}
