+++ src/main/java/com/fhsh/daitda/hubservice/hubinventory/application/dto/command/RestoreHubInventoryCommand.java
@@ -1,0 +1,14 @@
+package com.fhsh.daitda.hubservice.hubinventory.application.dto.command;
+
+import lombok.Builder;
+import lombok.Getter;
+
+import java.util.UUID;
+
+@Getter
+@Builder
+public class RestoreHubInventoryCommand {
+
+    private UUID hubInventoryId;
+    private Integer quantity;
+}
