+++ src/main/java/com/fhsh/daitda/hubservice/hubinventory/application/dto/command/UpdateHubInventoryCommand.java
@@ -1,0 +1,12 @@
+package com.fhsh.daitda.hubservice.hubinventory.application.dto.command;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+public class UpdateHubInventoryCommand {
+
+    private Integer quantity;
+
+}
