--- src/test/java/com/daitda/hubservice/HubServiceApplicationTests.java
@@ -1,13 +1,0 @@
-package com.daitda.hubservice;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class HubServiceApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
