+++ src/test/java/com/fhsh/daitda/hubservice/HubServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package com.fhsh.daitda.hubservice;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class HubServiceApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
